requirements.txt
# ===================================================
# AGIFORMER - Stable Requirements for NumPy 2.x
# ===================================================

# Core ML Framework
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0

# Foundational Libraries
numpy>=2.0.0
scipy>=1.13.0
scikit-learn>=1.4.0
pandas>=2.2.0

# Hugging Face Ecosystem
transformers>=4.41.0
accelerate>=0.30.0
datasets>=2.19.0
tokenizers>=0.19.0

# Utilities
einops>=0.8.0
pyyaml>=6.0
tqdm>=4.66.0
wandb>=0.17.0
Pillow>=10.0.0
requests>=2.31.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Optional High-Performance Libraries (Yoruma alındı)
# flash-attn>=2.5.8
